Rework Error handling to avoid race conditions
The way the probe screen handled errors was broken - it would be a race
between the probe screen and AXISUI calling LinuxCNC's error_channel.poll()
method and there was noi guarantee who would call it first. This results
in different behaviors depending on who calls it first.

The LinuxCNC docs explicitly call out that only a single error channel
reader should exist in any environment, so we simply shouldn't be reading
this ourselves.

Fixes #30